Update catppuccin-theme.el - delete expired/unpurchased domain
Hello from hacktoberfest,
The domain you are referring to (shttps://spec.catppuccintheme.com/) is available to everyone to purchase and can be taken over by an attacker. An attacker could exploit such a situation by registering the expired/unpurchased domain and setting up a phishing page that mimics your project.