docs: mention that `custom-theme-directory` must be set
By default, the value of `custom-theme-directory` in Emacs is "~/.emacs.d". So
if a user tried to run the code snippet in the documentation, he would be met
with an error saying that "catppuccin" couldn't be loaded.

If we tell the user to install the theme to "~/.emacs.d/themes", then we should
also tell him to set `custom-theme-directory` to that value.