fix: use `intern` over `make-symbol`
There are checks elsewhere in the code to test that
`catppuccin-flavor` is `'latte`. But if we create a symbol using
`make-symbol`, "its value and function definition are void, and its
property list is nil." (from the documentation)

`intern` instead finds a symbol that is already defined.

```elisp
(eq (make-symbol "latte") 'latte) ;; nil
(eq (intern "latte") 'latte)      ;; t
```